gateway-routers: Re-structure routers
