+++ pallets/connectors-gateway/routers/src/routers/mod.rs
@@ -1,0 +1,18 @@
+// Copyright 2021 Centrifuge Foundation (centrifuge.io).
+//
+// This file is part of the Centrifuge chain project.
+// Centrifuge is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version (see http://www.gnu.org/licenses).
+// Centrifuge is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+pub mod axelar_evm;
+pub mod axelar_xcm;
+pub mod ethereum_xcm;
+
+pub use axelar_evm::*;
+pub use axelar_xcm::*;
+pub use ethereum_xcm::*;
